all classes altered for integer

+++ src/main/java/com/habitalign/controllers/DaysOfWeekController.java
@@ -34,7 +34,7 @@ public class DaysOfWeekController {
 
 	@PutMapping("/updatedayofweek/{id}")
 	public ResponseEntity<Void> updateDaysOfWeek(@RequestBody DaysOfWeekActiveDTO daysOfWeekActiveDTO,
-			@PathVariable Long id) {
+			@PathVariable Integer id) {
 
 		daysOfWeekService.updateIsActiveDaysOfWeek(id, daysOfWeekActiveDTO);
 		return ResponseEntity.noContent().build();
